requirements.txt
streamlit==1.45.1
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
pillow==11.1.0
